app/requirements.txt
appdirs==1.4.4
apturl==0.5.2
astroid==2.4.2
awscli==1.18.69
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
boto3==1.17.12
botocore==1.20.12
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cloudscraper==1.2.52
colorama==0.4.3
command-not-found==0.3
cryptography==3.3.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gTTS==2.1.1
gTTS-token==1.1.4
httplib2==0.14.0
idna==2.8
iso8601==0.1.13
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jira==2.0.0
jmespath==0.9.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
m3u8==0.8.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
openshot-qt==2.4.3
pandas==1.2.2
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
playsound==1.2.2
protobuf==3.6.1
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
pymongo==3.11.2
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.15.0
soupsieve==2.0.1
systemd-python==234
testresources==2.0.1
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.2
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.31
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
